LF-13874: gpu: drm: imx-drm-core: Fix type mismatch warning in dev_err
Addresses a type mismatch warning encountered when using the dev_err()
with the `ARRAY_SIZE` macro on 32-bit compilers.

The `ARRAY_SIZE` macro can return a ‘unsigned int’ on 32-bit systems,
and return a `long unsigned int` on 64-bit system.
while dev_err() expects an `long int` for the format specifier.

This commit explicitly casts `ARRAY_SIZE(dpu)` to `u32` to ensure
compatibility with both 32-bit and 64-bit compilers.

+++ drivers/gpu/drm/imx/imx-drm-core.c
@@ -261,8 +261,9 @@ static void add_dpu_bliteng_components(struct device *dev,
 			found = false;
 		} else {
 			if (num_dpu >= ARRAY_SIZE(dpu)) {
-				dev_err(dev, "The number of found dpu is greater than max [%ld].\n",
-					ARRAY_SIZE(dpu));
+				dev_err(dev, "The number of found dpu is greater than max [%u].\n",
+					(u32)ARRAY_SIZE(dpu));
+
 				of_node_put(parent);
 				of_node_put(port);
 				break;
